Move code to create connection objects from sockets into separate module.
